% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinosaurs.R
\docType{data}
\name{dinasours}
\alias{dinasours}
\alias{dinosaurs}
\title{Dinosaur species}
\format{
A data frame with 20 observations on the following 4 variables.
\describe{
\item{name}{Name of species.}
\item{ecosystem}{Ecosystem the species lives in.}
\item{vore}{Vore of species: \code{carnivore}, \code{herbivore}, or \code{pescavore}.}
\item{name_mening}{Meaning of species name.}
}
}
\source{
Various
}
\usage{
dinosaurs
}
\description{
Numerous variables on ___ species of dinosaurs collected from various resources.
}
\examples{

library(ggplot2)

# Finishing times
ggplot(dinosaurs, aes(x = vore)) +
  geom_bar() +
  labs(
    title = "What dinosaurs eat",
    x = "Vore",
    y = "Count"
    )

}
\keyword{datasets}
