% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dinosaurs.R
\docType{data}
\name{dinasours}
\alias{dinasours}
\alias{dinosaurs}
\title{Dinosaur species}
\format{
A data frame with 20 observations on the following 4 variables.
\describe{
\item{name}{Name of species.}
\item{ecosystem}{Ecosystem the species lives in.}
\item{vore}{Vore of species: \code{carnivore}, \code{herbivore}, or \code{pescavore}.}
\item{name_mening}{Meaning of species name.}
}
}
\source{
Various, mostly Wikipedia.
}
\usage{
dinosaurs
}
\description{
Numerous variables on species of dinosaurs collected from various resources.
}
\examples{

library(ggplot2)

# What dinosaurs eat
ggplot(dinosaurs, aes(x = vore, fill = vore)) +
geom_bar(show.legend = FALSE) +
  labs(
    title = "What dinosaurs eat 🦖",
    x = "Vore",
    y = "Count"
  ) +
  scale_fill_manual(values = c("firebrick1", "forestgreen")) +
  theme_minimal()

# Where dinosaurs live
ggplot(dinosaurs, aes(y = ecosystem, fill = ecosystem)) +
geom_bar(show.legend = FALSE) +
  labs(
    title = "Where dinosaurs live 🌎",
    y = "Ecosystem",
    x = "Count"
  ) +
  scale_fill_manual(values = c("chocolate3", "darkseagreen2", "deepskyblue3")) +
  theme_minimal()
}
\keyword{datasets}
